Minor demo fixes
git-svn-id: svn://svn.hafro.is/jminor/trunk@17677 51816742-9f83-410d-85ed-e7bcf7042953

+++ demos/chinook/src/main/java/org/jminor/framework/demos/chinook/domain/impl/ChinookImpl.java
@@ -245,7 +245,7 @@ public final class ChinookImpl extends Domain implements Chinook {
             columnProperty(INVOICE_TOTAL, Types.DECIMAL, "Total")
                     .setMaximumFractionDigits(2)
                     .setHidden(true),
-            subqueryProperty(INVOICE_TOTAL_SUB, Types.DOUBLE, "Calculated total", INVOICE_TOTAL_SUBQUERY)
+            subqueryProperty(INVOICE_TOTAL_SUB, Types.DECIMAL, "Calculated total", INVOICE_TOTAL_SUBQUERY)
                     .setMaximumFractionDigits(2))
             .setKeyGenerator(automaticKeyGenerator("chinook.invoice"))
             .setOrderBy(orderBy().ascending(INVOICE_CUSTOMERID).descending(INVOICE_INVOICEDATE))
